api/v2/requirements.txt
# sudo apt-get install libpq-dev

django
djangorestframework
markdown
django-filter
django-cors-headers
psycopg2
elastic-apm
django-elasticsearch-dsl
python-logstash
djangorestframework-simplejwt